ISSU-1 Creates PlayerCollapse
