+++ src/PlayerCollapse/PlayerCollapse.css
@@ -1,0 +1,14 @@
+.PlayerCollapse {
+  display: flex;
+  justify-content: space-around;
+}
+
+.hidden-player {
+  opacity: 0;
+  pointer-events: none;
+  height: 0 !important;
+}
+
+.eye-icon {
+  margin-left: 5px;
+}
